track_dolphin: extern u8 pauseMenuGetState (renderGlows 90.01->90.63%)
The file-scope `extern int pauseMenuGetState` overrode the canonical u8
return (header + tricky.c def), so == 0 guards compiled to signed cmpwi.
Restoring u8 gives target's clrlwi r0,r3,24; cmplwi r0,0 (#11).

(cherry picked from commit 668217dd5d9acd853b7bb08e5cbe5a4c3617ec2f)

+++ src/main/track_dolphin.c
@@ -3904,7 +3904,7 @@ extern void skyBuildSunModelMatrix(f32 * out);
 extern void Camera_ProjectWorldPointWithOffset(f32 x, f32 y, f32 z, f32 offset, f32* outX, f32* outY, f32* outZ);
 extern void Camera_NdcToScreen(f32 x, f32 y, f32 z, int* ox, int* oy, int* oz);
 extern int depthReadRequestPoll(int x, int y, void* p);
-extern int pauseMenuGetState(void);
+extern u8 pauseMenuGetState(void);
 extern void* fn_8008912C(void);
 extern void _gxSetTevColor2(int r, int g, int b, int a);
 extern int sSynthFadeUnit;
